Move LogMetaPayload to separate file
